add new page with empty components
