+++ src/editor/components/landingPage/cta.tsx
@@ -1,0 +1,9 @@
+interface CTAProps {
+  props?: any
+}
+
+export const CTA = ( { props }: CTAProps ): JSX.Element => (
+  <>
+    <h2>Call to Action</h2>
+  </>
+)
