+++ src/editor/components/landingPage/features.tsx
@@ -1,0 +1,9 @@
+interface FeaturesProps {
+  props?: any
+}
+
+export const Features = ( { props }: FeaturesProps ): JSX.Element => (
+  <>
+    <h2>Features</h2>
+  </>
+)
