+++ src/editor/components/landingPage/footer.tsx
@@ -1,0 +1,9 @@
+interface FooterProps {
+  props?: any
+}
+
+export const Footer = ( { props }: FooterProps ): JSX.Element => (
+  <>
+    <h2>Footer </h2>
+  </>
+)
