+++ src/editor/components/landingPage/hero.tsx
@@ -1,0 +1,9 @@
+interface HeroProps {
+  props?: any
+}
+
+export const Hero = ( { props }: HeroProps ): JSX.Element => (
+  <>
+    <h2>Hero</h2>
+  </>
+)
