+++ src/editor/components/landingPage/pricing.tsx
@@ -1,0 +1,9 @@
+interface PricingProps {
+  props?: any
+}
+
+export const Pricing = ( { props }: PricingProps ): JSX.Element => (
+  <>
+    <h2>Pricing</h2>
+  </>
+)
