+++ src/editor/components/list/list-item.tsx
@@ -1,5 +1,3 @@
-import React from 'react'
-
 interface ListItemProps {
   children: any
 }
